# docs/swagger.yaml
basePath: /
definitions:
  accountant.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  accountant.GetOperatorResponse:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      operators:
        items:
          $ref: '#/definitions/accountant.Operator'
        type: array
      page:
        type: integer
      total_count:
        type: integer
    type: object
  accountant.Operator:
    properties:
      is_active:
        type: boolean
      park_no:
        type: string
      total_payment:
        type: number
      user_id:
        type: integer
      username:
        type: string
    type: object
  accountant.ResultById:
    properties:
      end_at:
        type: string
      opertor_id:
        type: integer
      start_at:
        type: string
      total_payment:
        type: integer
    type: object
  camera.CameraType:
    enum:
    - inside
    - outside
    type: string
    x-enum-varnames:
    - Inside
    - Outside
  camera.Cameras:
    properties:
      id:
        type: integer
      name:
        type: string
      type:
        $ref: '#/definitions/camera.CameraType'
    type: object
...
host: 192.168.100.7:3000
info:
  contact: {}
  title: Airline REST API
paths:
  /api/v1/accountant:
    get:
      consumes:
      - application/json
      description: Get list of operators with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 5
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/accountant.GetOperatorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/accountant.ErrorResponse'
      summary: Get list of operators
      tags:
      - Accountant
  /api/v1/accountant/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves payments for a specific operator within their login/logout
        periods
      parameters:
      - description: Operator ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/accountant.ResultById'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/accountant.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/accountant.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/accountant.ErrorResponse'
      summary: Get payments for a specific operator by ID
      tags:
      - Accountant
...
  /api/v1/cameras:
    get:
      description: Retrieves a list of cameras from the database with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/camera.Cameras'
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get cameras with pagination
      tags:
      - Cameras
    post:
      consumes:
      - application/json
      description: Creates a new camera in the database with validation for camera
        type
      parameters:
      - description: Camera data
        in: body
        name: camera
        required: true
        schema:
          $ref: '#/definitions/camera.Cameras'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/camera.Cameras'
        "400":
          description: Invalid camera type
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a new camera
      tags:
      - Cameras
  /api/v1/cameras/{id}:
    delete:
      description: Deletes the camera from the database using its unique ID
      parameters:
      - description: Camera ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Camera deleted successfully
          schema:
            type: string
        "404":
          description: Camera not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete a camera by ID
      tags:
      - Cameras
    get:
      description: Retrieves the camera from the database using its unique ID
      parameters:
      - description: Camera ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/camera.Cameras'
        "404":
          description: Camera not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get a camera by ID
      tags:
      - Cameras
    put:
      consumes:
      - application/json
      description: Updates the camera data in the database using its unique ID
      parameters:
      - description: Camera ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated camera data
        in: body
        name: camera
        required: true
        schema:
          $ref: '#/definitions/camera.Cameras'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/camera.Cameras'
        "400":
          description: Invalid camera data
          schema:
            type: string
        "404":
          description: Camera not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Update a camera by ID
      tags:
      - Cameras
...
swagger: "2.0"